// frontend/src/api/generated/model/jobBatchResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * crossbill API
 * OpenAPI spec version: 0.1.0
 */

export interface JobBatchResponse {
  id: number;
  batch_type: string;
  reference_id: string;
  total_jobs: number;
  completed_jobs: number;
  failed_jobs: number;
  status: string;
  created_at: string;
  updated_at: string;
}